// src/Application/ApiCalls/generated/models/InterventionPlanningDTO.ts
export interface InterventionPlanningDTO {
    /**
     * 
     * @type {string}
     * @memberof InterventionPlanningDTO
     */
    noIntervention?: string | null;
    /**
     * 
     * @type {string}
     * @memberof InterventionPlanningDTO
     */
    codeImmeuble?: string | null;
    /**
     * 
     * @type {string}
     * @memberof InterventionPlanningDTO
     */
    adressImmeuble?: string | null;
    /**
     * 
     * @type {string}
     * @memberof InterventionPlanningDTO
     */
    designation?: string | null;
    /**
     * 
     * @type {string}
     * @memberof InterventionPlanningDTO
     */
    datePrevu?: string | null;
    /**
     * 
     * @type {string}
     * @memberof InterventionPlanningDTO
     */
    latitude?: string | null;
    /**
     * 
     * @type {string}
     * @memberof InterventionPlanningDTO
     */
    longitude?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof InterventionPlanningDTO
     */
    __urgence?: boolean;
    /**
     * 
     * @type {InterventionCodeEtatDTO}
     * @memberof InterventionPlanningDTO
     */
    codeEtat?: InterventionCodeEtatDTO;
}