// src/Application/ApiCalls/generated/models/PlanningDTOOutput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AXECIEL.GMAO.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { InterventionPlanningDTO } from './InterventionPlanningDTO';
import {
    InterventionPlanningDTOFromJSON,
    InterventionPlanningDTOFromJSONTyped,
    InterventionPlanningDTOToJSON,
    InterventionPlanningDTOToJSONTyped,
} from './InterventionPlanningDTO';

/**
 * 
 * @export
 * @interface PlanningDTOOutput
 */
export interface PlanningDTOOutput {
    /**
     * 
     * @type {Array<InterventionPlanningDTO>}
     * @memberof PlanningDTOOutput
     */
    interventionJour?: Array<InterventionPlanningDTO> | null;
    /**
     * 
     * @type {Array<InterventionPlanningDTO>}
     * @memberof PlanningDTOOutput
     */
    interventionEnAttente?: Array<InterventionPlanningDTO> | null;
    /**
     * 
     * @type {Array<InterventionPlanningDTO>}
     * @memberof PlanningDTOOutput
     */
    historiqueIntervention?: Array<InterventionPlanningDTO> | null;
}

/**
 * Check if a given object implements the PlanningDTOOutput interface.
 */
export function instanceOfPlanningDTOOutput(value: object): value is PlanningDTOOutput {
    return true;
}

export function PlanningDTOOutputFromJSON(json: any): PlanningDTOOutput {
    return PlanningDTOOutputFromJSONTyped(json, false);
}

export function PlanningDTOOutputFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlanningDTOOutput {
    if (json == null) {
        return json;
    }
    return {
        
        'interventionJour': json['interventionJour'] == null ? undefined : ((json['interventionJour'] as Array<any>).map(InterventionPlanningDTOFromJSON)),
        'interventionEnAttente': json['interventionEnAttente'] == null ? undefined : ((json['interventionEnAttente'] as Array<any>).map(InterventionPlanningDTOFromJSON)),
        'historiqueIntervention': json['historiqueIntervention'] == null ? undefined : ((json['historiqueIntervention'] as Array<any>).map(InterventionPlanningDTOFromJSON)),
    };
}

export function PlanningDTOOutputToJSON(json: any): PlanningDTOOutput {
    return PlanningDTOOutputToJSONTyped(json, false);
}

export function PlanningDTOOutputToJSONTyped(value?: PlanningDTOOutput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'interventionJour': value['interventionJour'] == null ? undefined : ((value['interventionJour'] as Array<any>).map(InterventionPlanningDTOToJSON)),
        'interventionEnAttente': value['interventionEnAttente'] == null ? undefined : ((value['interventionEnAttente'] as Array<any>).map(InterventionPlanningDTOToJSON)),
        'historiqueIntervention': value['historiqueIntervention'] == null ? undefined : ((value['historiqueIntervention'] as Array<any>).map(InterventionPlanningDTOToJSON)),
    };
}